If DISK is set use ext4 file else use tmpfs

+++ entrypoint.sh
@@ -3,9 +3,11 @@
 #start sshd
 /etc/init.d/ssh start
 
-# Create the ext4 volume image
-dd if=/dev/zero of=/var/tmp/docker.img bs=1 count=0 seek=${DISK}
-mkfs.ext4 /var/tmp/docker.img
+# Create the ext4 volume image if DISK is set
+if [[ -n "${DISK}" ]] ; then
+    dd if=/dev/zero of=/var/tmp/docker.img bs=1 count=0 seek=${DISK}
+    mkfs.ext4 /var/tmp/docker.img
+fi
 
 #start the uml kernel with docker inside
 TMPDIR=/dev /sbin/start-stop-daemon --start --chuid `whoami` --chdir $PWD --background --make-pidfile --pidfile /tmp/kernel.pid --exec /linux/linux -- \
